updates to deploy to heroku

+++ src/app/Mycomponents/addto-cart/addto-cart.component.ts
@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
+import { MatBottomSheet } from '@angular/material/bottom-sheet';
 import { AuthService } from 'src/app/Services/auth.service';
+import { MysheetComponent } from '../mysheet/mysheet.component';
 
 export interface data {
   title: string;
   price: string;
+  qty: number;
 }
 
 @Component({
@@ -12,40 +15,78 @@ export interface data {
   styleUrls: ['./addto-cart.component.css']
 })
 export class AddtoCartComponent implements OnInit {
-  abc = [{'name':'abc'}]
   cartitems = [];
   localItem: string;
+  len: number;
+  cart_length
 
-  constructor(private auth: AuthService) { 
+  constructor(private auth: AuthService, private bottomSheet: MatBottomSheet) {
     this.localItem = localStorage.getItem("cartitem");
-    if (this.localItem == null){
-      this.cartitems=[];
+    if (this.localItem == null) {
+      this.cartitems = [];
+
     }
-    else{
+    else {
       this.cartitems = JSON.parse(this.localItem);
     }
   }
 
-  ngOnInit() : void{
-    this.auth.getMsg().subscribe((item:data) => {
+  ngOnInit(): void {
+    this.auth.getMsg().subscribe((item: data) => {
       console.log(item)
-
-      this.cartitems.push({
-        productName: item.title,
-        price: item.price
-      })
-
-      console.log(this.cartitems);
-      localStorage.setItem("cartitem", JSON.stringify(this.cartitems));
-      
-      
+      this.addProductToCart(item)
     })
-    console.log(this.cartitems);
   }
 
-  deleteitem(){
+  addProductToCart(item: data) {
+
+    let productExist = false;
+
+    this.localItem = localStorage.getItem("cartitem");
+    this.localItem = JSON.parse(this.localItem);
+
+    for (let i in this.cartitems) {
+      if (this.cartitems[i].productName === item.title) {
+        console.log("INCREASE QUANTITY")
+        this.cartitems[i].qty++
+
+        localStorage.setItem("cartitem", JSON.stringify(this.cartitems));
+        productExist = true
+        break;
+      }
+    }
+
+    if (!productExist) {
+
+      console.log("PUSH")
+      this.cartitems.push({
+        productName: item.title,
+        price: item.price,
+        qty: 1,
+      })
+
+      localStorage.setItem("cartitem", JSON.stringify(this.cartitems));
+    }
+    
+
+  }
+
+  deleteitem() {
     this.cartitems.splice(0, 1);
     localStorage.setItem("cartitem", JSON.stringify(this.cartitems));
+    
+    // getting data from local storage
+    this.localItem = localStorage.getItem("cartitem");
+
+    // parsing localstorage data
+    this.localItem = JSON.parse(this.localItem)
+    
+    this.cart_length = this.localItem.length
+    this.auth.changeDataSub(this.cart_length)
+  }
+
+  openBottomSheet(){
+    this.bottomSheet.open(MysheetComponent)
   }
